Move away from macros for both new and edit forms
This increases maintainability of the code.
