+++ lib/jumubase_web/live/internal/performance_live/edit.ex
@@ -1,12 +1,58 @@
 defmodule JumubaseWeb.Internal.PerformanceLive.Edit do
   use Phoenix.LiveView
-  use JumubaseWeb.PerformanceLive, :edit
   import Jumubase.Gettext
   import JumubaseWeb.PerformanceController, only: [normalize_params: 1]
   import JumubaseWeb.PerformanceLive.Helpers
+
+  import JumubaseWeb.PerformanceLive.Edit,
+    only: [
+      prepare: 2,
+      change: 2,
+      add_appearance: 1,
+      remove_appearance: 2,
+      add_piece: 1,
+      remove_piece: 2
+    ]
+
   alias Ecto.Changeset
   alias Jumubase.Showtime
   alias JumubaseWeb.Router.Helpers, as: Routes
+
+  def render(assigns) do
+    JumubaseWeb.PerformanceView.render("live_form.html", assigns)
+  end
+
+  def mount(_params, assigns, socket) do
+    {:ok, prepare(socket, assigns)}
+  end
+
+  def handle_event("change", %{"performance" => attrs}, socket) do
+    {:noreply, change(socket, attrs)}
+  end
+
+  def handle_event("add-appearance", _, socket) do
+    {:noreply, add_appearance(socket)}
+  end
+
+  def handle_event("remove-appearance", %{"index" => index}, socket) do
+    {:noreply, remove_appearance(socket, parse_id(index))}
+  end
+
+  def handle_event("toggle-appearance-panel", %{"index" => index}, socket) do
+    {:noreply, toggle_appearance_panel(socket, parse_id(index))}
+  end
+
+  def handle_event("add-piece", _, socket) do
+    {:noreply, add_piece(socket)}
+  end
+
+  def handle_event("remove-piece", %{"index" => index}, socket) do
+    {:noreply, remove_piece(socket, parse_id(index))}
+  end
+
+  def handle_event("toggle-piece-panel", %{"index" => index}, socket) do
+    {:noreply, toggle_piece_panel(socket, parse_id(index))}
+  end
 
   def handle_event("submit", %{"performance" => params}, socket) do
     contest = socket.assigns.contest
